Add files via upload
The LunarLander-v2 Box2D environment created by Oleg Klimov is out of date as of October 2025. Therefore, we have used version v3 for this task. The LunarLander-v3 environment is a classic rocket trajectory optimization problem. It has a multi-dimensional state vector containing (x, y) positional coordinates, velocities, angles, and angular velocities of a spacecraft, which is supposed to land between two flags. The agent is allowed to take discrete actions, i.e., selectively turn the spacecraft engines on or off. For Task 1, we formulated a simplified discrete-state discrete-action Markov Decision Process (MDP) that captures the essential dynamics of the lunar landing problem while remaining computationally tractable for value iteration.